ALWAYS store currency as DecimalField, not FloatField or you’re setting yourself up for trouble in the future

+++ corehq/apps/smsbillables/models.py
@@ -29,7 +29,7 @@ class SmsGatewayFee(models.Model):
     Once an SmsFee is created, it cannot be modified.
     """
     criteria = models.ForeignKey(SmsGatewayFeeCriteria, on_delete=models.PROTECT)
-    amount = models.FloatField(default=0.0)
+    amount = models.DecimalField(default=0.0, max_digits=7, decimal_places=4)
     currency = models.ForeignKey(accounting.Currency, on_delete=models.PROTECT)
     date_created = models.DateField(auto_now_add=True)
 
@@ -56,7 +56,7 @@ class SmsUsageFee(models.Model):
     Once an SmsUsageFee is created, it cannot be modified.
     """
     criteria = models.ForeignKey(SmsUsageFeeCriteria, on_delete=models.PROTECT)
-    amount = models.FloatField(default=0.0)
+    amount = models.DecimalField(default=0.0, max_digits=7, decimal_places=4)
     date_created = models.DateField(auto_now_add=True)
